+++ src/FScript/CliArgs.fs
@@ -3,7 +3,7 @@ module CliArgs
 open Argu
 
 type CliArgs =
-    | [<MainCommand; ExactlyOnce; Last>] Script of path: string
+    | [<MainCommand; Last>] Script of path: string
     | [<AltCommandLine("-r")>] Root of path: string
     interface IArgParserTemplate with
         member this.Usage =